LayoutTests/fast/js/toString-exception-expected.txt

This test checks for regression against 7343: REGRESSION: fast/js/toString-overrides.html fails when run multiple times.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caught is true
PASS successfullyParsed is true

TEST COMPLETE
